chore: refactor rate button with modal to single component, refactor claim button
