--- src/Components/ClaimButton/styles.scss
@@ -1,10 +1,0 @@
-.claim-button {
-  animation: pulse 1s infinite cubic-bezier(0.25, 0, 0, 1);
-  box-shadow: 0 0 0 0 white;
-}
-
-@keyframes pulse {
-  to {
-    box-shadow: 0 0 0 6px rgba(255, 255, 255, 0);
-  }
-}
