+++ src/Components/RateButtonWithModal/styles.scss
