--- src/Components/RatingModal/types.ts
@@ -1,4 +1,0 @@
-export type RatingModalPropTypes = {
-  open: boolean;
-  onClose: () => void;
-};
