+++ src/Components/index.ts
@@ -6,12 +6,9 @@ export { default as SpreadBalancePad } from "./SpreadBalancePad/SpreadBalancePad
 export { default as DailyBonusModal } from "./DailyBonusModal/DailyBonusModal";
 export { default as Card } from "./Card/Card";
 export { default as CardsGroup } from "./CardsGroup/CardsGroup";
-export { default as ClaimButton } from "./ClaimButton/ClaimButton";
 export { default as BuyButton } from "./BuyButton/BuyButton";
 export { default as Settings } from "./Settings/Settings";
 export { default as BackgroundLayer } from "./BackgroundLayer/BackgroundLayer";
-export { default as RatingModal } from "./RatingModal/RatingModal";
-export { default as FeedbackForm } from "./FeedbackForm/FeedbackForm";
 export { default as MarkdownToHTML } from "./MarkdownToHTML/MarkdownToHTML";
 export { default as Preloader } from "./Preloader/Preloader";
 export { default as ChangeLanguageModal } from "./ChangeLanguageModal/ChangeLanguageModal";
